+++ tsclientlib/examples/channeltree.rs
@@ -88,7 +88,7 @@ async fn real_main() -> Result<()> {
 	// Wait some time
 	let mut events = con.events().try_filter(|_| future::ready(false));
 	tokio::select! {
-		_ = &mut time::delay_for(Duration::from_secs(1)) => {}
+		_ = time::sleep(Duration::from_secs(1)) => {}
 		_ = events.next() => {
 			bail!("Disconnected");
 		}
@@ -112,7 +112,7 @@ async fn real_main() -> Result<()> {
 	// Wait some time
 	let mut events = con.events().try_filter(|_| future::ready(false));
 	tokio::select! {
-		_ = &mut time::delay_for(Duration::from_secs(3)) => {}
+		_ = time::sleep(Duration::from_secs(3)) => {}
 		_ = events.next() => {
 			bail!("Disconnected");
 		}