+++ tsclientlib/examples/sync.rs
@@ -66,7 +66,7 @@ async fn real_main() -> Result<()> {
 	.await??;
 
 	// Wait some time
-	time::delay_for(Duration::from_secs(1)).await;
+	time::sleep(Duration::from_secs(1)).await;
 
 	// Disconnect
 	con.disconnect(DisconnectOptions::new()).await?;